# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "E-commerce",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "host": "http://localhost:8000/",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get All Products",
                "operationId": "testing",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add Product",
                "operationId": "ApiSaveProduct",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "stock",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "discount",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show Product By Product Id",
                "operationId": "ApiShowProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update Product",
                "operationId": "ApiUpdateProduct",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "stock",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "discount",
                        "in": "formData",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product By Product Id",
                "operationId": "ApiDeleteProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {}
}